Move more classes to namespace
